did a small refactor
implemented providers for domain classes
so that we can properly add the caching layer
inbetween.

refactored packages to be more fitting to
screaming architecture.
