+++ domain/src/main/java/be/xplore/githubmetrics/domain/providers/ports/WorkflowRunsProvider.java
@@ -1,0 +1,10 @@
+package be.xplore.githubmetrics.domain.providers.ports;
+
+import be.xplore.githubmetrics.domain.domain.WorkflowRun;
+
+import java.util.List;
+
+public interface WorkflowRunsProvider {
+
+    List<WorkflowRun> getLastDaysWorkflowRuns(String repository);
+}
