+++ domain/src/main/java/be/xplore/githubmetrics/domain/schedulers/ports/WorkflowRunsUseCase.java
@@ -1,4 +1,4 @@
-package be.xplore.githubmetrics.domain.usecases;
+package be.xplore.githubmetrics.domain.schedulers.ports;
 
 public interface WorkflowRunsUseCase {
     void retrieveAndExportWorkflowRuns();
